ci(e2e): recalibrate timings, split deps-graph e2e, raise parallelism to 32
