+++ .circleci/config.yml
@@ -768,7 +768,7 @@ jobs:
     environment:
       BITSRC_ENV: stg
       BIT_FEATURES: cloud-importer-v2
-    parallelism: 25
+    parallelism: 32
     steps:
       - e2e_test_cmd
 
@@ -778,7 +778,7 @@ jobs:
     environment:
       BITSRC_ENV: stg
       BIT_FEATURES: cloud-importer-v2
-    parallelism: 25
+    parallelism: 32
     steps:
       - attach_workspace:
           at: ./
@@ -819,7 +819,7 @@ jobs:
     environment:
       BITSRC_ENV: stg
       BIT_FEATURES: cloud-importer-v2
-    parallelism: 25
+    parallelism: 32
     steps:
       - attach_workspace:
           at: ./
